+++ bin/unprovision.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
-USAGE="Usage: unprovision.sh [-h service-broker-url]  -i instance-id"
+USAGE="Usage: unprovision.sh [-h service-broker-url]  -i instance-id [-u username] [-c password]"
 
 HOST=127.0.0.1:5001
 
-while getopts ":h:i:" opt; do
+while getopts ":h:i:u:c:" opt; do
   case $opt in
     h)
 	HOST=$OPTARG
@@ -11,6 +11,12 @@ while getopts ":h:i:" opt; do
     i)
 	INSTANCE_ID=$OPTARG
       ;;
+    u)
+	SERVICE_BROKER_USERNAME=$OPTARG
+	;;
+    c)
+	SERVICE_BROKER_PASSWORD=$OPTARG
+	;;
     \?) 
       echo $USAGE >&2 
       echo "Invalid option: -$OPTARG" >&2
@@ -25,10 +31,14 @@ if [ -z "$INSTANCE_ID" ] ; then
 	exit 1;
 fi
 
-USER=$(cat config/aws-rds-service-broker.json | jq -r ".credentials.authUser")
-PWD=$(cat config/aws-rds-service-broker.json | jq -r ".credentials.authPassword")
+if [ -z "$SERVICE_BROKER_USERNAME" -o -z "$SERVICE_BROKER_PASSWORD" ] ; then
+	echo $USAGE >&2
+	echo ERROR: missing SERVICE_BROKER_USERNAME or SERVICE_BROKER_PASSWORD >&2
+	exit 1;
+fi
 
-curl http://$USER:$PWD@$HOST/v2/service_instances/$INSTANCE_ID \
+
+curl -s -H 'x-broker-api-version: 2.4' http://$SERVICE_BROKER_USERNAME:$SERVICE_BROKER_PASSWORD@$HOST/v2/service_instances/$INSTANCE_ID \
 	 -H "Content-Type: application/json" \
 	-X DELETE \
 	--data '{
